# openapi.yaml
openapi: 3.0.0
info:
  title: Test Automation API
  description: API for managing test scripts and ECID information
  version: 1.0.0
servers:
  - url: http://localhost:8196
    description: Stress Rack server

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Secret-Key
    IpWhitelist:
      type: http
      scheme: bearer
      bearerFormat: ip
  responses:
    Unauthorized:
      description: Invalid or missing secret key
      content:
        application/json:
          example:
            error: "Invalid or missing secret key"
    NotFound:
      description: Resource not found
      content:
        application/json:
          example:
            status: "error"
            message: "Resource not found"

security:
  - ApiKeyAuth: []
  - IpWhitelist: []

paths:
  /api/preload:
    get:
      summary: 预请求信息
      description: 预请求信息，加快前端响应
      responses:
        200:
          description: 成功
          content:
            application/json:
              example:
                status: "success"

  /api/ecids:
    get:
      summary: 获取 ECID 列表
      description: 获取所有 ECID
      responses:
        200:
          description: 成功
          content:
            application/json:
              example:
                ["0010441438002936", "0010441438002675", "0010441438002676", "0010441438002677", "0010441438002678", "0010441438002679"]

  /api/testing_ecids:
    get:
      summary: 获取测试用 ECID 列表
      description: 获取测试 ECID
      responses:
        200:
          description: 成功
          content:
            application/json:
              example:
                ["0010441438002675", "0010441438003321", "0010441438003322", "0010441438003323"]

  /api/detail_info:
    post:
      summary: 获取 ECID 详细信息
      description: 根据 ECID 获取设备的详细信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ecids:
                  type: array
                  items:
                    type: string
              example:
                ecids: ["0010441438002936"]
      responses:
        200:
          description: 成功获取 ECID 详细信息
          content:
            application/json:
              example:
                0010441438002936:
                  battery: "100"
                  cg_vendor: "SDC"
                  temperature: "25.000000"
                  check_timestamp: 1776196934
                  config: "3PDFC5BW"
                  serial_number: "F2029KV7R1"
                  sw_vers: "24A62611c"
                  unit_number: "025178"
                0010441438002676:
                  battery: "86"
                  cg_vendor: "LGD"
                  temperature: "30.000000"
                  check_timestamp: 1776196934
                  config: "5G13S"
                  serial_number: "A8WQ7MUQ2D"
                  sw_vers: "24A62611c"
                  unit_number: "000312"
                0010441438002677:
                  battery: "10"
                  cg_vendor: "SDC"
                  temperature: "42.000000"
                  check_timestamp: 1776196934
                  config: "2KS2A" 
                  serial_number: "QM28EFW2NS"
                  sw_vers: "24A62612h"
                  unit_number: "00276"
                0010441438002678:
                  battery: "100"
                  cg_vendor: "SDC"
                  temperature: "45.000000"
                  check_timestamp: 1776196934
                  config: "1G40AS"
                  serial_number: "F2029KV7R1"
                  sw_vers: "24A62611d"
                  unit_number: "024178"
                0010441438002679:
                  battery: "100"
                  cg_vendor: "LGD"
                  temperature: "13.000000"
                  check_timestamp: 1776196934
                  config: "3PDFCA"
                  serial_number: "ER6F7TYQZM"
                  sw_vers: "24A62612f"
                  unit_number: "025138"


  /api/logs:
    post:
      summary: 获取测试日志流
      description: 流式获取日志信息（Server-Sent Events）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ecids:
                  type: array
                  items:
                    type: string
              example:
                ecids: ["0010441438002936"]
      responses:
        200:
          description: 日志流
          content:
            text/event-stream:
              example: |
                RUNNING TEST ['reference']
                VALID TEST. CONTINUING
                INFO: project_version sha1 = 3bedb055cca8df6a2cf7e53a85c70a13afdda38f
                INFO: project: XXX sha: 3bedb055cca8df6a2cf7e53a85c70a13afdda38f not in valid_sha1_map. Update sha1_map ...................
                INFO: project_version sha1 = 1ee687bc0b0198b3d2a43869990eaac403fce0a5
                INFO: line 1
                INFO: line 2
                INFO: line 3
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: line ...
                INFO: project_version: I3
                ...

  /api/script_info:
    get:
      summary: 获取脚本信息
      description: 获取所有测试脚本的基本信息
      responses:
        200:
          description: 成功
          content:
            application/json:
              example:
                DOE343:
                  path: "/Users/Mac/Desktop/DOE/DOE343_FactoryCoexXXX_V319_S"
                  status: false
                DOE344:
                  path: "/Users/Mac/Desktop/DOE/DOE344_FactoryCoexXXX_V319_S"
                  status: true
                DOE345:
                  path: "/Users/Mac/Desktop/DOE/DOE345_FactoryCoexXXX_V319_S"
                  status: false
                DOE3436:
                  path: "/Users/Mac/Desktop/DOE/DOE346_FactoryCoexXXX_V319_S"
                  status: false
                fix_essex:
                  path: "/Users/Mac/Desktop/Al_EVT_Factory/FactoryCoexXXX_V319_fix_essex"
                  status: true
                fix_pinacle:
                  path: "/Users/Mac/Desktop/Al_EVT_Factory/FactoryCoexXXX_V319_fix_pinacle"
                  status: true


  /api/script_methods:
    post:
      summary: 获取脚本的方法列表
      description: 根据脚本名获取该脚本中定义的所有方法名称
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                script_name:
                  type: string
                  description: 脚本名称（不包含路径）
                script_path:
                  type: string
                  description: 脚本路径（可选）
              required:
                - script_name
            example:
              script_name: "common"
      responses:
        200:
          description: 成功返回脚本的方法列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  methods:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
                    description: 脚本名 => 方法列表的映射
              example:
                status: "success"
                methods:
                  fix_essex:
                    - "a173aod"
                    - "a173camera"
                    - "a173peridot"
                    - "a173sensors"
                    - "a173ttw"
                    - "aod"
                    - "b390aod"
                    - "b390camera"
                    - ...
                  fix_pinacle:
                    - "a173aod"
                    - "a173camera"
                    - "a173peridot"
                    - "a173sensors"
                    - "a173ttw"
                    - "aod"
                    - "b390aod"
                    - "b390camera"
                    - ...
                  DOE344:
                    - "a173aod"
                    - "a173camera"
                    - "a173peridot"
                    - "a173sensors"
                    - "a173ttw"
                    - "aod"
                    - "b390aod"
                    - "b390camera"
                    - ...

  /api/add_test_script:
    post:
      summary: 添加测试脚本
      description: 添加新的测试脚本及其路径
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                script_name:
                  type: string
                script_path:
                  type: string
              example:
                script_name: "new_script"
                script_path: "/path/to/script.py"
      responses:
        200:
          description: 成功
          content:
            application/json:
              example:
                status: "success"
                added script: "/path/to/script.py"

  /api/scan_test_script:
    get:
      summary: 扫描主机的测试脚本
      description: 扫描主机的测试脚本
      responses:
        200:
          description: 成功
          content:
            application/json:
              example:
                DOE343_FactoryCoexXXX_V319_S:
                  - "/Users/Mac/Desktop/DOE/DOE343_FactoryCoexXXX_V319_S"
                DOE445_FactoryCoexXXX_V405_C:
                  - "/Users/Mac/Desktop/DOE/DOE445_FactoryCoexXXX_V405_C"
                  - "/Users/Mac/Desktop/EVT_Factory/DOE445_FactoryCoexXXX_V405_C"


  /api/remove_test_script:
    post:
      summary: 移除测试脚本
      description: 根据脚本名或路径移除测试脚本
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                script_name:
                  type: string
                script_path:
                  type: string
              example:
                script_name: "old_script"
      responses:
        200:
          description: 成功
          content:
            application/json:
              example:
                status: "success"
                removed script: "old_script"

  /api/run_test_script:
    post:
      summary: 运行测试脚本
      description: 运行指定的测试脚本
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                test_method:
                  type: string
                script_name:
                  type: string
                units:
                  type: array
                  items:
                    type: string
              example:
                test_method: "run_test"
                script_name: "test_script"
                units: ["ECID1", "ECID2"]
      responses:
        200:
          description: 成功
          content:
            application/json:
              example:
                status: "success"
                test_method: "run_test"
                units: ["0010441438002936"]

  /api/force_quit:
    post:
      summary: 强制退出当前测试脚本
      description: 强制退出当前测试脚本
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                units:
                  type: array
                  items:
                    type: string
              example:
                units: ["ECID1", "ECID2"]
      responses:
        200:
          description: 成功
          content:
            application/json:
              example:
                units: ["0010441438002675", "0010441438003321"]

  /api/discharge_battery:
    post:
      summary: 执行机台放电程序
      description: 执行机台放电程序
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ecids:
                  type: array
                target_power:
                  type: integer
              example:
                ecids: ["ECID1", "ECID2"]
                target_power: 5
      responses:
        200:
          description: 成功
          content:
            application/json:
              example:
                status: "success"
                discharge_result: "Start discharging battery"

  /api/send_ssh_cmd:
    post:
      summary: 发送允许的ssh指令
      description: 发送允许的ssh指令
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ecids:
                  type: array
                cmd:
                  type: string
              example:
                ecids: [ "ECID1", "ECID2" ]
                target_power: "reboot"
      responses:
        200:
          description: 成功
          content:
            application/json:
              example:
                status: "success"
                discharge_result: "Success send reboot"

  /api/get_cmd_list:
    get:
      summary: 获取允许的CMD
      description: 获取允许的CMD
      responses:
        200:
          description: 成功
          content:
            application/json:
              example:
                ["halt","reboot","os_app","diags","iboot","kill_colortest","renew_units"]